Separate metrics-related sqg code into separate module
